Finish the adventure loop portion of the game.
These changes finish up the adventure loop portion of the game.
There is now a way to win the game by simply choosing the second
option and leaving each of the three areas. Also, there's now a
reset button at the end that allows a player to start over after
winning the game. It will also be reused for when the player loses.