Add basic battle system and uncovered restart bug.
Added a basic battle system that randomly chooses 1 of 2 enemies
depending on location and allows the player to attack them, use magic,
or attempt to run. Once the ememy is dead or they successfuly run,
they get the win screen. The enemies do not yet fight back.

I also discovered a bug with the restart button that causes you to
instantly win games for some reason after you start a new match despite
the enemy hp showing correctly, and the only function determining a win
being after the enemies HP reaches 0 or lower.