Try battle turn system.
Tried a battle turn system on this branch. It failed but I'm keeping
it around to see if someone else can't troubleshoot it. Enemy seems
to take multiple compounding turns after hitting the 'OK' button to
progress the battle. I ~might~ know what the problem is. It may be
that after hitting the fight button once, it sort of 'unlocks' the code
inside permenantly so that every subsequent hit of the 'OK' button runs
the enemyTurn() function again, and without the need to hit the fight
button first. If that is the case, I have no clue how to fix it.